pyproject.toml
[project]
name = "vibe-bot"
version = "0.1.0"
description = "Telegram OCR bot for expense tracking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram==3.27.0",
    "aiosqlite==0.22.1",
    "google-genai==1.70.0",
    "openpyxl==3.1.5",
    "pandas==2.3.3",
    "pillow==12.2.0",
    "pydantic==2.12.5",
    "pydantic-settings>=2.6.0",
    "python-dotenv==1.2.2",
    "sqlalchemy==2.0.49",
    "loguru>=0.7.2",
]

[project.scripts]
vibe-bot = "vibe.bot:run"

[project.optional-dependencies]
dev = [
    "ruff>=0.12.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["vibe"]
